V1.79(2014-09-13 18:09) HostID变成一个自增的int64值与SOCKET句柄无关 	id范围很多兆，链接达到很多兆亿之后，id将转一圈回来，需要时间大约几十年 	所以当某个链接持续几十年在线的极端情况下，才可能导致2个链接ID相同

+++ source/frame/netserver/NetHost.cpp
@@ -56,7 +56,7 @@ bool NetHost::IsServer()
 	return m_pConnect->IsServer();
 }
 
-int NetHost::ID()
+int64 NetHost::ID()
 {
 	if ( NULL == m_pConnect ) return -1;
 	return m_pConnect->GetID();
